requirements.txt
aiosqlite==0.20.0
annotated-types==0.7.0
iso8601==2.1.0
pydantic==2.10.1
pydantic_core==2.27.1
pygame==2.6.1
pygame-menu==4.4.8
pyperclip==1.9.0
pypika-tortoise==0.2.2
pytz==2024.2
tortoise-orm==0.22.0
typing_extensions==4.12.2
